Controller class and Service class added

+++ src/main/java/com/example/PhoneBook/repository/PhoneRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.PhoneBook.model.Directory;
 
 @Repository
-public interface PhoneRepository extends JpaRepository<Directory,Long>{
+public interface PhoneRepository extends JpaRepository<Directory,Integer>{
 
 }
